Logo: Remove unused component Header/Logo
